chore: review Card components
