+++ libs/catalog/src/components/CardGrid/Card.tsx
@@ -9,15 +9,13 @@ import {
   MouseEvent,
   useCallback,
   useEffect,
-  useLayoutEffect,
-  useRef,
   useState,
 } from 'react';
 import type { CardProps } from '../../models/card-props';
 import { EntityHeader } from '../EntityHeader/EntityHeader';
 import { FolderPath } from '../FolderPath/FolderPath';
 import { StarToggleButton } from '../StarToggleButton/StarToggleButton';
-import { TopicTag } from '../TopicTag/TopicTag';
+import { TopicsLine } from '../TopicTag/TopicTag';
 import styles from './CardGrid.module.scss';
 
 /** Card for the Browse grid with highlighted search text and optional featured styling. */
@@ -55,38 +53,6 @@ export const Card: FC<CardProps> = ({
   useEffect(() => {
     setIsStarred(initialIsStarred ?? false);
   }, [initialIsStarred]);
-  const [visibleCount, setVisibleCount] = useState(item.topics.length);
-  const topicsRef = useRef<HTMLDivElement>(null);
-
-  const topicsKey = item.topics.join('\0');
-  useLayoutEffect(() => {
-    const container = topicsRef.current;
-    if (!container || item.topics.length === 0) {
-      setVisibleCount(item.topics.length);
-      return;
-    }
-
-    const children = Array.from(container.children) as HTMLElement[];
-    if (children.length === 0) return;
-
-    const firstTop = children[0].offsetTop;
-    const rowHeight = children[0].offsetHeight;
-    const limitTop = firstTop + rowHeight * 2;
-
-    let cutoff = children.length;
-    for (let i = 0; i < children.length; i++) {
-      if (children[i].offsetTop >= limitTop) {
-        cutoff = i;
-        break;
-      }
-    }
-
-    // If there is overflow, reduce by one to leave room for the "+N" badge
-    setVisibleCount(
-      cutoff < children.length ? Math.max(0, cutoff - 1) : children.length,
-    );
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [topicsKey, 2]);
 
   const handleToggle = (e: MouseEvent<HTMLElement>) => {
     e.stopPropagation();
@@ -109,8 +75,6 @@ export const Card: FC<CardProps> = ({
     [onClick, item],
   );
 
-  const overflow = item.topics.length - visibleCount;
-
   return (
     <div
       {...(onClick
@@ -122,7 +86,7 @@ export const Card: FC<CardProps> = ({
           }
         : {})}
       className={mergeClasses(
-        'relative box-border flex cursor-pointer flex-col gap-2.5 rounded-[6px] border p-[17px] transition-transform duration-150 ease-out hover:-translate-y-[3px]',
+        'relative box-border flex cursor-pointer flex-col gap-2 rounded-[6px] border p-4 transition-transform duration-150 ease-out hover:-translate-y-[4px]',
         styles.card,
         item.isFeatured ? styles.featuredCard : undefined,
         className,
@@ -146,15 +110,15 @@ export const Card: FC<CardProps> = ({
         <Highlight text={item.description} query={query} />
       </p>
 
-      <div className="mt-auto">
-        <div ref={topicsRef} className="flex flex-wrap gap-1.5">
-          {item.topics.slice(0, visibleCount).map((p) => (
-            <TopicTag key={p} label={p} />
-          ))}
-          {overflow > 0 && <TopicTag label={`+${overflow}`} />}
-        </div>
+      <div className="mt-auto flex flex-col gap-4">
+        <TopicsLine topics={item.topics} />
 
-        <div className="mt-4 flex items-center justify-between border-t border-secondary pt-2">
+        <div
+          className={mergeClasses(
+            'flex items-center justify-between border-t pt-2',
+            styles.cardFooter,
+          )}
+        >
           <FolderPath segments={item.folder} />
           <StarToggleButton isStarred={isStarred} onClick={handleToggle} />
         </div>
