+++ libs/catalog/src/components/CardGrid/CardGrid.tsx
@@ -41,7 +41,7 @@ export const CardGrid: FC<CardGridProps> = memo(
             <div
               key={i}
               className={mergeClasses(
-                'rounded-[6px] border p-[17px]',
+                'rounded-[6px] border p-4',
                 styles.skeletonCard,
               )}
               style={{ height: CARD_HEIGHT }}
