+++ libs/catalog/src/components/Details/DetailsPanel.module.scss
@@ -20,10 +20,6 @@
   padding-bottom: 6px;
 }
 
-.starFilledIcon {
-  color: var(--text-warning-icon, #eec840);
-}
-
 .content {
   scrollbar-width: thin;
   scrollbar-color: var(--stroke-secondary, #242c42) transparent;
