+++ libs/catalog/src/components/ListView/ListView.module.scss
@@ -1,7 +1,3 @@
 .secondaryText {
   color: var(--cat-list-secondary-text, var(--text-secondary, #9fa6bd));
 }
-
-.starFilledIcon {
-  color: var(--cat-card-star-filled, var(--text-warning-icon, #eec840));
-}
